requirements.txt
alembic==1.8.1
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.5
pyasn1==0.4.8
pydantic==1.10.2
python-decouple==3.6
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
